requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fake-useragent==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
praw==7.8.1
prawcore==2.4.0
pydantic==2.11.3
pydantic_core==2.33.1
pyotp==2.9.0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
twscrape==0.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
win32_setctime==1.2.0
